Move check for duplicate request
Move it to sidr_req handling.

This check affects future SIDR requests. IF a forward already exist then
another SIDR request cannot be issues. This is in particular an issue
if the source QP cannot be determined. Then this check will hold off
new SIDRs from going through until the app has started sending on
the QP. Which then allows us to determine the sourcd QP.

Signed-off-by: Christoph Lameter <cl@linux.com>